added permissions
code still isnt the best or the nicest but it does work again
and added some permissions

+++ resources/views/components/author-buttons.blade.php
@@ -1,12 +1,16 @@
-@props(['user'=> false, 'edit_href'=> '#','delete_href' => '#'])
+@props(['user'=> false, 'edit_href'=> '#','delete_href' => '#' , 'task' => false])
 
 @if($user)
     <div class="flex items-center justify-evenly w-full">
-        <a href="{{$edit_href}}">
-            <x-button name="edit"/>
-        </a>
-        <x-form action="{{$delete_href}}" method="DELETE">
-            <x-button name="delete"/>
-        </x-form>
+        @if(!$task || hasPermission('task_edit'))
+            <a href="{{$edit_href}}">
+                <x-button name="edit"/>
+            </a>
+        @endif
+        @if(!$task || hasPermission('task_delete'))
+            <x-form action="{{$delete_href}}" method="DELETE">
+                <x-button name="delete"/>
+            </x-form>
+        @endif
     </div>
 @endif
